[Decomp][Spell] Cohesion split: extract SpellEffectHealPower handlers to SpellEffectHealPower.cpp
